refactor: clarify source module helpers
